+++ axum/src/json.rs
@@ -15,9 +15,16 @@ use std::ops::{Deref, DerefMut};
 /// JSON Extractor / Response.
 ///
 /// When used as an extractor, it can deserialize request bodies into some type that
-/// implements [`serde::Deserialize`]. If the request body cannot be parsed, or it does not contain
-/// the `Content-Type: application/json` header, it will reject the request and return a
-/// `400 Bad Request` response.
+/// implements [`serde::Deserialize`]. The request will be rejected (and a [`JsonRejection`] will
+/// be returned) if:
+///
+/// - The request doesn't have a `Content-Type: application/json` (or similar) header.
+/// - The body doesn't contain syntactically valid JSON.
+/// - The body contains syntactically valid JSON but it couldn't be deserialized into the target
+/// type.
+/// - Buffering the request body fails.
+///
+/// See [`JsonRejection`] for more details.
 ///
 /// # Extractor example
 ///
@@ -99,7 +106,27 @@ where
         if json_content_type(req) {
             let bytes = Bytes::from_request(req).await?;
 
-            let value = serde_json::from_slice(&bytes).map_err(InvalidJsonBody::from_err)?;
+            let value = match serde_json::from_slice(&bytes) {
+                Ok(value) => value,
+                Err(err) => {
+                    let rejection = match err.classify() {
+                        serde_json::error::Category::Data => JsonDataError::from_err(err).into(),
+                        serde_json::error::Category::Syntax | serde_json::error::Category::Eof => {
+                            JsonSyntaxError::from_err(err).into()
+                        }
+                        serde_json::error::Category::Io => {
+                            if cfg!(debug_assertions) {
+                                // we don't use `serde_json::from_reader` and instead always buffer
+                                // bodies first, so we shouldn't encounter any IO errors
+                                unreachable!()
+                            } else {
+                                JsonSyntaxError::from_err(err).into()
+                            }
+                        }
+                    };
+                    return Err(rejection);
+                }
+            };
 
             Ok(Json(value))
         } else {
@@ -243,5 +270,20 @@ mod tests {
         assert!(valid_json_content_type("application/json;charset=utf-8").await);
         assert!(valid_json_content_type("application/cloudevents+json").await);
         assert!(!valid_json_content_type("text/json").await);
+    }
+
+    #[tokio::test]
+    async fn invalid_json_syntax() {
+        let app = Router::new().route("/", post(|_: Json<serde_json::Value>| async {}));
+
+        let client = TestClient::new(app);
+        let res = client
+            .post("/")
+            .body("{")
+            .header("content-type", "application/json")
+            .send()
+            .await;
+
+        assert_eq!(res.status(), StatusCode::BAD_REQUEST);
     }
 }