+++ axum/src/routing/tests/nest.rs
@@ -1,8 +1,10 @@
-use tower_http::services::ServeDir;
-
 use super::*;
-use crate::{body::boxed, extract::Extension};
+use crate::{
+    body::boxed,
+    extract::{Extension, MatchedPath},
+};
 use std::collections::HashMap;
+use tower_http::services::ServeDir;
 
 #[tokio::test]
 async fn nesting_apps() {
@@ -346,23 +348,81 @@ async fn outer_middleware_still_see_whole_url() {
     assert_eq!(client.get("/one/two").send().await.text().await, "/one/two");
 }
 
+#[tokio::test]
+async fn nest_at_capture() {
+    let api_routes = Router::new()
+        .route(
+            "/:b",
+            get(|Path((a, b)): Path<(String, String)>| async move { format!("a={} b={}", a, b) }),
+        )
+        .boxed_clone();
+
+    let app = Router::new().nest("/:a", api_routes);
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar").send().await;
+    assert_eq!(res.status(), StatusCode::OK);
+    assert_eq!(res.text().await, "a=foo b=bar");
+}
+
 macro_rules! nested_route_test {
+    (
+        $name:ident,
+        nest = $nested_path:literal,
+        route = $route_path:literal,
+        expected = $expected_path:literal,
+        opaque_redirect = $opaque_redirect:expr $(,)?
+    ) => {
+        #[tokio::test]
+        async fn $name() {
+            let inner = Router::new().route(
+                $route_path,
+                get(|matched: MatchedPath| async move { matched.as_str().to_owned() }),
+            );
+            let app = Router::new().nest($nested_path, inner);
+            let client = TestClient::new(app);
+            let res = client.get($expected_path).send().await;
+            let status = res.status();
+            let matched_path = res.text().await;
+            assert_eq!(status, StatusCode::OK, "Router");
+
+            let inner = Router::new()
+                .route(
+                    $route_path,
+                    get(|matched: MatchedPath| async move { matched.as_str().to_owned() }),
+                )
+                .boxed_clone();
+            let app = Router::new().nest($nested_path, inner);
+            let client = TestClient::new(app);
+            let res = client.get($expected_path).send().await;
+            if $opaque_redirect {
+                assert_eq!(res.status(), StatusCode::PERMANENT_REDIRECT, "opaque");
+
+                let location = res.headers()[http::header::LOCATION].to_str().unwrap();
+                let res = client.get(location).send().await;
+                assert_eq!(res.status(), StatusCode::OK, "opaque with redirect");
+                assert_eq!(res.text().await, location);
+            } else {
+                assert_eq!(res.status(), StatusCode::OK, "opaque");
+                assert_eq!(res.text().await, matched_path);
+            }
+        }
+    };
+
     (
         $name:ident,
         nest = $nested_path:literal,
         route = $route_path:literal,
         expected = $expected_path:literal $(,)?
     ) => {
-        #[tokio::test]
-        async fn $name() {
-            let inner = Router::new().route($route_path, get(|| async {}));
-            let app = Router::new().nest($nested_path, inner);
-            let client = TestClient::new(app);
-            assert_eq!(
-                client.get($expected_path).send().await.status(),
-                StatusCode::OK
-            );
-        }
+        nested_route_test!(
+            $name,
+            nest = $nested_path,
+            route = $route_path,
+            expected = $expected_path,
+            opaque_redirect = false,
+        );
     };
 }
 
@@ -373,7 +433,22 @@ nested_route_test!(nest_3, nest = "", route = "/a/", expected = "/a/");
 nested_route_test!(nest_4, nest = "/", route = "/", expected = "/");
 nested_route_test!(nest_5, nest = "/", route = "/a", expected = "/a");
 nested_route_test!(nest_6, nest = "/", route = "/a/", expected = "/a/");
-nested_route_test!(nest_7, nest = "/a", route = "/", expected = "/a");
+
+// This case is different for opaque services.
+//
+// The internal route becomes `/a/*__private__axum_nest_tail_param` which, according to matchit
+// doesn't match `/a`. However matchit detects that a route for `/a/` exists and so it issues a
+// redirect to `/a/`, which ends up calling the inner route as expected.
+//
+// So while the behavior isn't identical, the outcome is the same
+nested_route_test!(
+    nest_7,
+    nest = "/a",
+    route = "/",
+    expected = "/a",
+    opaque_redirect = true,
+);
+
 nested_route_test!(nest_8, nest = "/a", route = "/a", expected = "/a/a");
 nested_route_test!(nest_9, nest = "/a", route = "/a/", expected = "/a/a/");
 nested_route_test!(nest_11, nest = "/a/", route = "/", expected = "/a/");