+++ axum-core/CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- **added:** Add `IntoResponseParts` trait ([#797])
+- **added:** Add `IntoResponseParts` trait which allows defining custom response
+  types for adding headers or extensions to responses ([#797])
 - **added:** Implement `IntoResponse` for `bytes::BytesMut` and `bytes::Chain<T, U>` ([#767])
 - **breaking:** Using `HeaderMap` as an extractor will no longer remove the headers and thus
   they'll still be accessible to other extractors, such as `axum::extract::Json`. Instead
@@ -36,11 +37,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `<Request as FromRequest>::Rejection` is now `BodyAlreadyExtracted`.
     - `<http::request::Parts as FromRequest>::Rejection` is now `Infallible`.
     - `ExtensionsAlreadyExtracted` has been removed.
+- **breaking:** `RequestParts::body_mut` now returns `&mut Option<B>` so the
+  body can be swapped ([#869])
 
 [#698]: https://github.com/tokio-rs/axum/pull/698
 [#699]: https://github.com/tokio-rs/axum/pull/699
 [#767]: https://github.com/tokio-rs/axum/pull/767
 [#797]: https://github.com/tokio-rs/axum/pull/797
+[#869]: https://github.com/tokio-rs/axum/pull/869
 
 # 0.1.1 (06. December, 2021)